marked event and dir create and list.files() fix
marked_event does not work for GOA files because there are less of those than other haul events.
added in a line to run_method() make a directory that was not previously there.
added some if statements for listing out files in the instances they don't exist.

+++ R/make_cast_sections.R
@@ -162,10 +162,14 @@ make_cast_sections <- function(haul_metadata_path = list.files(paste0(getwd(), "
   print("Binning upcast and downcast data")
   system(command = paste0("sbebatch ", getwd(), "/", binavg_bat, " ", getwd(), " ", xmlcon_file))
   
+  if(length(list.files(path = here::here("cnv"), pattern = "_uuu_", full.names = TRUE)) >= 1){
   upcast_files <- list.files(path = here::here("cnv"), pattern = "_uuu_", full.names = TRUE)
   file.remove(upcast_files[grep(pattern = "downcast", x = upcast_files)])
+  }
   
+  if(length(list.files(path = here::here("cnv"), pattern = "_ddd_", full.names = TRUE)) >= 1){
   downcast_files <- list.files(path = here::here("cnv"), pattern = "_ddd_", full.names = TRUE)
   file.remove(downcast_files[grep(pattern = "upcast", x = downcast_files)])
+  }
   
 }
